Cycles: Limit number of open file handles used by the texture cache
Processes have a limit on the number of open file handles, and the OIIO
image input also comes with its own memory overhead. This is shared
across all Cycles instances.

Currently this is hardcoded to 128, we should add a better user heuristic
and perhaps manual override for usage of Cycles outside Blender.

The Attic scene randomly fails to render on the CPU without this.